update building menu's

+++ Assets/Code/Runtime/_Contracts/Configs/Buildings/BuildingConfig.cs
@@ -4,7 +4,7 @@ namespace SustainTheStrain._Contracts.Configs.Buildings
 {
     public abstract class BuildingConfig : ScriptableObject
     {
-        [field: SerializeField, Min(1)] public int Level { get; private set; } = 1;
+        [field: SerializeField, Min(1)] public byte Level { get; private set; } = 1;
         [field: SerializeField, Min(0)] public int Price { get; private set; }
         [field: SerializeField, Min(0)] public int Compensation { get; private set; }
